pyproject.toml
[project]
name = "st-camera-opencv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.6",
    "opencv-python-headless>=4.12.0.88",
    "pillow>=11.3.0",
    "plotly>=6.3.0",
    "streamlit>=1.49.1",
]
